+++ common/src/main/java/fr/tathan/exoconfig/ExoConfig.java
@@ -2,17 +2,16 @@ package fr.tathan.exoconfig;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.google.gson.Strictness;
 import com.google.gson.ToNumberPolicy;
-import commonnetwork.api.Dispatcher;
 import fr.tathan.exoconfig.client.ExodusClient;
+import fr.tathan.exoconfig.common.Config;
 import fr.tathan.exoconfig.common.infos.CommentTypeAdapter;
 import fr.tathan.exoconfig.common.loader.ConfigsRegistry;
-import fr.tathan.exoconfig.common.network.NetworkRegistry;
+import fr.tathan.exoconfig.common.network.ProxyExclusionAdapterFactory;
 import fr.tathan.exoconfig.common.network.SyncConfigPacket;
-import fr.tathan.exoconfig.common.Config;
 import fr.tathan.exoconfig.common.types.ConfigTypesRegistry;
 import fr.tathan.exoconfig.common.types.RangedNumber;
-import fr.tathan.exoconfig.common.network.ProxyExclusionAdapterFactory;
 import fr.tathan.exoconfig.common.types.ResourceLocationType;
 import fr.tathan.exoconfig.platform.PlatformHelper;
 import net.minecraft.resources.ResourceLocation;
@@ -26,7 +25,7 @@ public final class ExoConfig {
     public static final Logger LOG = LoggerFactory.getLogger(MOD_ID);
     private static final GsonBuilder GSON = new GsonBuilder()
             .setPrettyPrinting()
-            .setLenient()
+            .setStrictness(Strictness.LENIENT)
             .registerTypeAdapterFactory(new CommentTypeAdapter())
             .registerTypeAdapterFactory(new ProxyExclusionAdapterFactory())
             .setObjectToNumberStrategy(ToNumberPolicy.LONG_OR_DOUBLE);
@@ -35,7 +34,6 @@ public final class ExoConfig {
 
     public static void init() {
         ExodusClient.init();
-        NetworkRegistry.init();
 
         ConfigTypesRegistry.register(RangedNumber.class, () -> new RangedNumber(0, 100, 50));
         ConfigTypesRegistry.register(ResourceLocation.class, ResourceLocationType::serialize, ResourceLocationType::deserialize);
@@ -58,7 +56,7 @@ public final class ExoConfig {
         if (joined) {
             ConfigsRegistry.getInstance().getConfigs().forEach((key, config) -> {
                 if (config.isSyncable()) {
-                    Dispatcher.sendToClient(new SyncConfigPacket(key, getGson().toJson(config)), player);
+                    PlatformHelper.sendToClient(new SyncConfigPacket(key, getGson().toJson(config)), player);
                 }
             });
         }