+++ common/src/main/java/fr/tathan/exoconfig/client/ExodusClient.java
@@ -5,7 +5,6 @@ import fr.tathan.exoconfig.client.components.EnumButton;
 import fr.tathan.exoconfig.client.components.RangedOption;
 import fr.tathan.exoconfig.client.components.StateButton;
 import fr.tathan.exoconfig.common.types.RangedNumber;
-import fr.tathan.exoconfig.common.types.ResourceLocationType;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.components.Tooltip;
 import net.minecraft.network.chat.Component;