+++ common/src/main/java/fr/tathan/exoconfig/client/components/StateButton.java
@@ -1,6 +1,5 @@
 package fr.tathan.exoconfig.client.components;
 
-import com.mojang.blaze3d.systems.RenderSystem;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.AbstractButton;