+++ common/src/main/java/fr/tathan/exoconfig/common/types/ConfigType.java
@@ -4,7 +4,6 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonSerializationContext;
-import fr.tathan.exoconfig.ExoConfig;
 
 import java.lang.reflect.Type;