+++ common/src/main/java/fr/tathan/exoconfig/platform/PlatformHelper.java
@@ -2,6 +2,8 @@ package fr.tathan.exoconfig.platform;
 
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
+import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
+import net.minecraft.server.level.ServerPlayer;
 
 import java.nio.file.Path;
 
@@ -14,6 +16,11 @@ public class PlatformHelper {
 
     @ExpectPlatform
     public static void registerConfigScreen(String modid, Object config) {
+        throw new AssertionError();
+    }
+
+    @ExpectPlatform
+    public static void sendToClient(CustomPacketPayload payload, ServerPlayer player) {
         throw new AssertionError();
     }
 }