+++ fabric/src/main/java/fr/tathan/exoconfig/fabric/client/ExoconfigFabricClient.java
@@ -1,10 +1,11 @@
 package fr.tathan.exoconfig.fabric.client;
 
+import fr.tathan.exoconfig.fabric.network.NetworkRegistry;
 import net.fabricmc.api.ClientModInitializer;
 
 public final class ExoconfigFabricClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
-        // This entrypoint is suitable for setting up client-specific logic, such as rendering.
+        NetworkRegistry.initClient();
     }
 }